Try to shrink diff
