+++ scripts/reset-database.ts
@@ -1,57 +1,30 @@
 import { config } from 'dotenv';
 import postgres from 'postgres';
-import { getHostUrl } from '../lib/databricks-host-utils';
+import { getDatabricksToken } from '../lib/auth/databricks-auth-node';
+import { getPostgresUrlFromEnv, getDatabaseConfigFromEnv, buildConnectionUrl } from '../lib/db/connection-core';
 
 config({ path: '.env.local' });
 
-async function getDatabricksToken() {
-  const clientId = process.env.DATABRICKS_CLIENT_ID;
-  const clientSecret = process.env.DATABRICKS_CLIENT_SECRET;
-
-  if (!clientId || !clientSecret) {
-    throw new Error('DATABRICKS_CLIENT_ID, DATABRICKS_CLIENT_SECRET, and DATABRICKS_HOST must be set');
-  }
-
-  const tokenUrl = `${getHostUrl()}/oidc/v1/token`;
-
-  const response = await fetch(tokenUrl, {
-    method: 'POST',
-    headers: {
-      Authorization: `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`,
-      'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    body: 'grant_type=client_credentials&scope=all-apis',
-  });
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    throw new Error(`Failed to get OAuth token: ${response.status} ${errorText}`);
-  }
-
-  const data = await response.json();
-  return data.access_token;
-}
-
 async function getConnectionUrl() {
-  if (process.env.POSTGRES_URL) {
-    return process.env.POSTGRES_URL;
+  // Use POSTGRES_URL if available
+  const postgresUrl = getPostgresUrlFromEnv();
+  if (postgresUrl) {
+    return postgresUrl;
+  }
+
+  // Build from components using shared utilities
+  const config = getDatabaseConfigFromEnv();
+  if (!config) {
+    throw new Error('Either POSTGRES_URL or PGHOST and PGDATABASE must be set');
   }
 
   const pgUser = process.env.PGUSER;
-  const pgHost = process.env.PGHOST;
-  const pgPort = process.env.PGPORT || '5432';
-  const pgDatabase = process.env.PGDATABASE;
-  const pgSSLMode = process.env.PGSSLMODE || 'require';
-
-  if (!pgUser || !pgHost || !pgDatabase) {
-    throw new Error('Either POSTGRES_URL or PGUSER, PGHOST, and PGDATABASE must be set');
+  if (!pgUser) {
+    throw new Error('PGUSER must be set for OAuth authentication');
   }
 
   const token = await getDatabricksToken();
-  const encodedUser = encodeURIComponent(pgUser);
-  const encodedPassword = encodeURIComponent(token);
-
-  return `postgresql://${encodedUser}:${encodedPassword}@${pgHost}:${pgPort}/${pgDatabase}?sslmode=${pgSSLMode}`;
+  return buildConnectionUrl(config, { username: pgUser, password: token });
 }
 
 async function resetDatabase() {
